+++ src/test/java/org/goafabric/calleeservice/extensions/TenantContextTest.java
@@ -33,7 +33,7 @@ class TenantContextTest {
 
     @Test
     void getUserNameFromUserINfo() {
-        var userInfo = "eyJwcmVmZXJyZWRfdXNlcm5hbWUiOiJqb2huIGRvZSIsImFsZyI6IkhTMjU2In0";
+        var userInfo = "eyJwcmVmZXJyZWRfdXNlcm5hbWUiOiJqb2huIGRvZSIsImFsZyI6IkhTMjU2In0"; //eyJnZW5kZXIiOiJVTktOT1dOIiwiYmlydGhkYXRlIjoiMTk4MC0wMS0wMSIsImVtYWlsX3ZlcmlmaWVkIjp0cnVlLCJ1cm46Z29hZmFicmljOmNsYWltczppbnN0aXR1dGlvbiI6W3siaWQiOiI4MmZhY2RkZS0xOTVjLTQ3OGUtODczYS0yM2MzNDBmZmYyZWQiLCJyb2xlcyI6WyJNRU1CRVIiLCJBRE1JTiJdLCJuYW1lIjoiVGVzdCJ9XSwibmFtZSI6IkpvbmggRG9lIiwicHJlZmVycmVkX3VzZXJuYW1lIjoiam9obi5kb2VAdW5rbm93bi5vcmciLCJlbWFpbCI6ImpvaG4uZG9lQHVua25vd24ub3JnIiwiYWxnIjoiSFMyNTYifQ
         TenantContext.setContext("42", "44", "user", userInfo);
         assertThat(TenantContext.getUserName()).isEqualTo("john doe");
     }
@@ -46,6 +46,5 @@ class TenantContextTest {
         assertThat(TenantContext.getUserName()).isEqualTo("user");
         assertThat(TenantContext.getAdapterHeaderMap()).isNotNull().isNotEmpty();
     }
-
 
 }