+++ src/main/java/org/goafabric/calleeservice/extensions/TenantContext.java
@@ -1,59 +1,68 @@
 package org.goafabric.calleeservice.extensions;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 
+import java.io.IOException;
+import java.util.Base64;
 import java.util.Map;
 
 public class TenantContext {
-    private static final ThreadLocal<TenantContextRecord> tenantContext = ThreadLocal.withInitial(() -> new TenantContextRecord(null, null, null));
-
     record TenantContextRecord(String tenantId, String organizationId, String userName) {
         public Map<String, String> toAdapterHeaderMap() {
             return Map.of("X-TenantId", tenantId, "X-OrganizationId", organizationId, "X-Auth-Request-Preferred-Username", userName);
         }
     }
 
+    private static final ThreadLocal<TenantContextRecord> CONTEXT =
+            ThreadLocal.withInitial(() -> new TenantContextRecord("0", "0", "anonymous"));
+
     public static void setContext(HttpServletRequest request) {
-        tenantContext.set(new TenantContextRecord(
-                request.getHeader("X-TenantId"),
-                request.getHeader("X-OrganizationId"),
-                request.getHeader("X-Auth-Request-Preferred-Username")));
+        setContext(request.getHeader("X-TenantId"), request.getHeader("X-OrganizationId"),
+                request.getHeader("X-Auth-Request-Preferred-Username"), request.getHeader("X-UserInfo"));
     }
 
-    static void setContext(TenantContextRecord tenantContextRecord) {
-        tenantContext.set(tenantContextRecord);
+    static void setContext(String tenantId, String organizationId, String userName, String userInfo) {
+        CONTEXT.set(new TenantContextRecord(
+                getValue(tenantId, "0"),
+                getValue(organizationId, "0"),
+                getValue(getUserNameFromUserInfo(userInfo), getValue(userName, "anonymous"))
+        ));
     }
 
     public static void removeContext() {
-        tenantContext.remove();
+        CONTEXT.remove();
     }
 
-    public static void setTenantId(String tenant) {
-        tenantContext.set(new TenantContextRecord(tenant, tenantContext.get().organizationId, tenantContext.get().userName));
+    private static String getValue(String value, String defaultValue) {
+        return value != null ? value : defaultValue;
     }
 
     public static String getTenantId() {
-        return tenantContext.get().tenantId() != null ? tenantContext.get().tenantId() : "0";
+        return CONTEXT.get().tenantId();
     }
 
     public static String getOrganizationId() {
-        return tenantContext.get().organizationId() != null ? tenantContext.get().organizationId() : "0";
+        return CONTEXT.get().organizationId();
     }
 
     public static String getUserName() {
-        return (getAuthentication() != null) && !(getAuthentication().getName().equals("anonymousUser")) ? getAuthentication().getName()
-                : tenantContext.get().userName != null ? tenantContext.get().userName : "anonymous";
+        return CONTEXT.get().userName();
     }
 
     public static Map<String, String> getAdapterHeaderMap() {
-        return tenantContext.get().toAdapterHeaderMap();
+        return CONTEXT.get().toAdapterHeaderMap();
     }
 
-    private static Authentication getAuthentication() {
-        return SecurityContextHolder.getContext().getAuthentication();
+    public static void setTenantId(String tenant) {
+        CONTEXT.set(new TenantContextRecord(tenant, CONTEXT.get().organizationId, CONTEXT.get().userName));
     }
 
-
+    private static String getUserNameFromUserInfo(String userInfo) {
+        try {
+            return userInfo != null ? (String) new ObjectMapper().readValue(Base64.getUrlDecoder().decode(userInfo), Map.class).get("preferred_username") : null;
+        } catch (IOException e) {
+            throw new IllegalStateException(e);
+        }
+    }
 }